src/requirements.txt
crewai[google-genai]==1.2.1
pydantic == 2.12.3
google-genai == 1.47.0
python-dotenv == 1.2.1
tavily-python == 0.7.12
streamlit == 1.30.0
pyttsx3 ==2.99
